pyproject.toml
[project]
name = "ai-blog-editor"
version = "0.1.0"
description = "Local RAG + style tools for writing blog posts in your own voice"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "ai-blog-editor" }]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pydantic>=2.7.4",
  "pyyaml>=6.0.2",
  "httpx>=0.27.0",
  "chromadb>=0.5.5",
  "python-frontmatter>=1.1.0",
  "markdown-it-py>=3.0.0",
  "python-dotenv>=1.0.1",
]

[project.scripts]
aiblog = "aiblog.cli:app"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["aiblog"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
